+++ braindecode/models/modules.py
@@ -25,6 +25,7 @@ from braindecode.models.functions import (
 )
 from braindecode.util import np_to_th
 from braindecode.models.eegminer import GeneralizedGaussianFilter
+from braindecode.models.parametrization import MaxNorm
 
 
 class Ensure4d(nn.Module):
@@ -343,28 +344,6 @@ class MaxNormLinear(nn.Linear):
         self._max_norm_val = max_norm_val
         self._eps = eps
         register_parametrization(self, "weight", MaxNorm(max_norm_val, eps))
-
-
-class MaxNorm(nn.Module):
-    def __init__(self, max_norm_val=2.0, eps=1e-5):
-        super().__init__()
-        self.max_norm_val = max_norm_val
-        self.eps = eps
-
-    def forward(self, X):
-        norm = X.norm(2, dim=0, keepdim=True)
-        denom = norm.clamp(min=self.max_norm_val / 2)
-        number = denom.clamp(max=self.max_norm_val)
-        return X * (number / (denom + self.eps))
-
-    def right_inverse(self, X):
-        # Assuming the forward scales X by a factor s,
-        # the right inverse would scale it back by 1/s.
-        norm = X.norm(2, dim=0, keepdim=True)
-        denom = norm.clamp(min=self.max_norm_val / 2)
-        number = denom.clamp(max=self.max_norm_val)
-        scale = number / (denom + self.eps)
-        return X / scale
 
 
 class LinearWithConstraint(nn.Linear):
